Added staged changes, updating changes in the ColorMonitor thread, and improved the logger.
Changes in the settings window are no longer directly updated in game.
Well, they weren't able to update without a restart until now. This allows
changes to be seen in game. The console is now a singleton so it is easier
to write to a log file. Which is really nice when something happens during
map load and you can't see the console.